rewrite descriptions
Rewrite the functions descriptions from c++ comments to markdown documentation

+++ math/exgcd.cpp
@@ -1,4 +1,4 @@
-int exgcd(int a, int b, long long &x, long long &y) {
+int exgcd(int a, int b, ll &x, ll &y) {
     if (b == 0) {
         x = 1, y = 0;
         return a;
